test: separate balance sync tests by component, remove duplicates, rename to convention :test_tube:
