+++ components/ledger/internal/bootstrap/balance_sync.worker_mt_property_test.go
@@ -1,0 +1,111 @@
+// Copyright (c) 2026 Lerian Studio. All rights reserved.
+// Use of this source code is governed by the Elastic License 2.0
+// that can be found in the LICENSE file.
+
+package bootstrap
+
+import (
+	"testing"
+	"testing/quick"
+
+	tmclient "github.com/LerianStudio/lib-commons/v4/commons/tenant-manager/client"
+	tmpostgres "github.com/LerianStudio/lib-commons/v4/commons/tenant-manager/postgres"
+	"github.com/LerianStudio/lib-commons/v4/commons/tenant-manager/tenantcache"
+	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/command"
+	"github.com/stretchr/testify/require"
+)
+
+// TestProperty_BalanceSyncWorker_PredicateEqualsEnabledAndPGManagerNonNil verifies
+// that isMTReady() on BalanceSyncWorker always equals the conjunction of
+// mtEnabled && pgManager != nil && tenantCache != nil, for any
+// combination of the three fields generated by quick.Check.
+func TestProperty_BalanceSyncWorker_PredicateEqualsEnabledAndPGManagerNonNil(t *testing.T) {
+	t.Parallel()
+
+	logger := newTestLogger()
+	useCase := &command.UseCase{}
+	tc, err := tmclient.NewClient("http://localhost:0", logger, tmclient.WithAllowInsecureHTTP(), tmclient.WithServiceAPIKey("test-api-key"))
+	require.NoError(t, err)
+	pgMgr := tmpostgres.NewManager(tc, "transaction", tmpostgres.WithLogger(logger))
+	cache := tenantcache.NewTenantCache()
+
+	property := func(enabled bool, hasPGManager bool, hasTenantCache bool) bool {
+		worker := NewBalanceSyncWorker(logger, useCase, BalanceSyncConfig{})
+		worker.mtEnabled = enabled
+
+		if hasPGManager {
+			worker.pgManager = pgMgr
+		} else {
+			worker.pgManager = nil
+		}
+
+		if hasTenantCache {
+			worker.tenantCache = cache
+		} else {
+			worker.tenantCache = nil
+		}
+
+		got := worker.isMTReady()
+		expected := enabled && hasPGManager && hasTenantCache
+
+		// Property: predicate result MUST equal (enabled && pgManager != nil && tenantCache != nil).
+		if got != expected {
+			return false
+		}
+
+		// Property: idempotent -- calling again yields the same result.
+		return worker.isMTReady() == got
+	}
+
+	err = quick.Check(property, &quick.Config{MaxCount: 200})
+	require.NoError(t, err,
+		"Property violated: BalanceSyncWorker.isMTReady() != (mtEnabled && pgManager != nil && tenantCache != nil)")
+}
+
+// TestProperty_NewBalanceSyncWorkerMT_PreservesBaseFields verifies that
+// the multi-tenant constructor preserves all base constructor fields AND correctly
+// sets multi-tenant fields.
+func TestProperty_NewBalanceSyncWorkerMT_PreservesBaseFields(t *testing.T) {
+	t.Parallel()
+
+	logger := newTestLogger()
+	useCase := &command.UseCase{}
+	tenantClient, err := tmclient.NewClient("http://localhost:0", logger, tmclient.WithAllowInsecureHTTP(), tmclient.WithServiceAPIKey("test-api-key"))
+	require.NoError(t, err)
+	pgMgr := tmpostgres.NewManager(tenantClient, "transaction", tmpostgres.WithLogger(logger))
+
+	cache := tenantcache.NewTenantCache()
+
+	property := func(enabled bool) bool {
+		worker := NewBalanceSyncWorkerMT(logger, useCase, BalanceSyncConfig{}, enabled, cache, pgMgr, "transaction")
+
+		// Property: constructor never returns nil.
+		if worker == nil {
+			return false
+		}
+
+		// Property: base fields preserved (referential equality).
+		if worker.useCase != useCase {
+			return false
+		}
+
+		// Property: multi-tenant fields set correctly.
+		if worker.mtEnabled != enabled {
+			return false
+		}
+
+		if worker.tenantCache != cache {
+			return false
+		}
+
+		if worker.pgManager != pgMgr {
+			return false
+		}
+
+		return true
+	}
+
+	err = quick.Check(property, &quick.Config{MaxCount: 100})
+	require.NoError(t, err,
+		"Property violated: NewBalanceSyncWorkerMT did not preserve base fields")
+}
