+++ components/ledger/internal/bootstrap/redis_consumer.worker_mt_property_test.go
@@ -4,31 +4,6 @@
 
 package bootstrap
 
-// =============================================================================
-// PROPERTY-BASED TESTS -- Multi-tenant worker/consumer domain invariants
-//
-// These tests verify that the domain invariants of the multi-tenant constructors
-// and predicates hold across hundreds of automatically-generated inputs using
-// testing/quick.Check.
-//
-// Invariants verified:
-//   1. Predicate correctness: isMTReady() / isMultiTenantReady() == (enabled && pgManager != nil)
-//      for any struct state, on both BalanceSyncWorker and RedisQueueConsumer.
-//   2. Constructor field preservation: NewBalanceSyncWorkerMT preserves
-//      base constructor fields AND sets multi-tenant fields correctly.
-//   3. Constructor field preservation: NewRedisQueueConsumerMultiTenant preserves
-//      base constructor fields AND sets multi-tenant fields correctly.
-//   4. maxWorkers default guard: if maxWorkers <= 0, it is clamped to 5.
-//   5. Nil-safety: no constructor or predicate ever panics for any combination
-//      of nil/non-nil pointers and bool/int values.
-//
-// Run with:
-//
-//	go test -run TestProperty -v -count=1 \
-//	    ./components/transaction/internal/bootstrap/
-//
-// =============================================================================
-
 import (
 	"testing"
 	"testing/quick"
@@ -41,61 +16,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TestProperty_BalanceSyncWorker_PredicateEqualsEnabledAndPGManagerNonNil verifies
-// that isMTReady() on BalanceSyncWorker always equals the conjunction of
-// mtEnabled && pgManager != nil && tenantCache != nil, for any
-// combination of the three fields generated by quick.Check.
-//
-// This guards against drift where the predicate logic diverges from its
-// documented contract (e.g., adding extra conditions or dropping one).
-func TestProperty_BalanceSyncWorker_PredicateEqualsEnabledAndPGManagerNonNil(t *testing.T) {
-	t.Parallel()
-
-	logger := newTestLogger()
-	useCase := &command.UseCase{}
-	tc, err := tmclient.NewClient("http://localhost:0", logger, tmclient.WithAllowInsecureHTTP(), tmclient.WithServiceAPIKey("test-api-key"))
-	require.NoError(t, err)
-	pgMgr := tmpostgres.NewManager(tc, "transaction", tmpostgres.WithLogger(logger))
-	cache := tenantcache.NewTenantCache()
-
-	property := func(enabled bool, hasPGManager bool, hasTenantCache bool) bool {
-		worker := NewBalanceSyncWorker(logger, useCase, BalanceSyncConfig{})
-		worker.mtEnabled = enabled
-
-		if hasPGManager {
-			worker.pgManager = pgMgr
-		} else {
-			worker.pgManager = nil
-		}
-
-		if hasTenantCache {
-			worker.tenantCache = cache
-		} else {
-			worker.tenantCache = nil
-		}
-
-		got := worker.isMTReady()
-		expected := enabled && hasPGManager && hasTenantCache
-
-		// Property: predicate result MUST equal (enabled && pgManager != nil && tenantCache != nil).
-		if got != expected {
-			return false
-		}
-
-		// Property: idempotent -- calling again yields the same result.
-		return worker.isMTReady() == got
-	}
-
-	err = quick.Check(property, &quick.Config{MaxCount: 200})
-	require.NoError(t, err,
-		"Property violated: BalanceSyncWorker.isMTReady() != (mtEnabled && pgManager != nil && tenantCache != nil)")
-}
-
 // TestProperty_RedisQueueConsumer_PredicateEqualsEnabledAndPGManagerNonNil verifies
 // the same predicate correctness invariant on RedisQueueConsumer.
-//
-// This ensures both structs share identical predicate semantics, preventing
-// divergence where one struct's predicate is updated but the other is not.
 func TestProperty_RedisQueueConsumer_PredicateEqualsEnabledAndPGManagerNonNil(t *testing.T) {
 	t.Parallel()
 
@@ -140,70 +62,9 @@ func TestProperty_RedisQueueConsumer_PredicateEqualsEnabledAndPGManagerNonNil(t 
 		"Property violated: RedisQueueConsumer.isMultiTenantReady() != (multiTenantEnabled && pgManager != nil && tenantCache != nil)")
 }
 
-// TestProperty_NewBalanceSyncWorkerMT_PreservesBaseFields verifies that
-// the multi-tenant constructor preserves all base constructor fields AND correctly
-// sets multi-tenant fields.
-//
-// Properties checked for every generated enabled (bool):
-//   - useCase is the same instance passed in
-//   - mtEnabled matches the input bool
-//   - tenantCache and pgManager are the exact instances passed in
-func TestProperty_NewBalanceSyncWorkerMT_PreservesBaseFields(t *testing.T) {
-	t.Parallel()
-
-	logger := newTestLogger()
-	useCase := &command.UseCase{}
-	tenantClient, err := tmclient.NewClient("http://localhost:0", logger, tmclient.WithAllowInsecureHTTP(), tmclient.WithServiceAPIKey("test-api-key"))
-	require.NoError(t, err)
-	pgMgr := tmpostgres.NewManager(tenantClient, "transaction", tmpostgres.WithLogger(logger))
-
-	cache := tenantcache.NewTenantCache()
-
-	property := func(enabled bool) bool {
-		worker := NewBalanceSyncWorkerMT(logger, useCase, BalanceSyncConfig{}, enabled, cache, pgMgr, "transaction")
-
-		// Property: constructor never returns nil.
-		if worker == nil {
-			return false
-		}
-
-		// Property: base fields preserved (referential equality).
-		if worker.useCase != useCase {
-			return false
-		}
-
-		// Property: multi-tenant fields set correctly.
-		if worker.mtEnabled != enabled {
-			return false
-		}
-
-		if worker.tenantCache != cache {
-			return false
-		}
-
-		if worker.pgManager != pgMgr {
-			return false
-		}
-
-		return true
-	}
-
-	err = quick.Check(property, &quick.Config{MaxCount: 100})
-	require.NoError(t, err,
-		"Property violated: NewBalanceSyncWorkerMT did not preserve base fields")
-}
-
 // TestProperty_NewRedisQueueConsumerMultiTenant_PreservesBaseFields verifies that
 // the multi-tenant constructor preserves all base constructor fields AND correctly
 // sets the multi-tenant fields.
-//
-// Properties checked for every generated enabled (bool) and hasPGManager (bool):
-//   - Logger is the same instance passed in
-//   - TransactionHandler is the same value passed in
-//   - multiTenantEnabled matches the input
-//   - tenantClient is the exact instance passed in
-//   - pgManager is set or nil depending on hasPGManager
-//   - isMultiTenantReady() is consistent with the field combination
 func TestProperty_NewRedisQueueConsumerMultiTenant_PreservesBaseFields(t *testing.T) {
 	t.Parallel()
 
@@ -263,12 +124,7 @@ func TestProperty_NewRedisQueueConsumerMultiTenant_PreservesBaseFields(t *testin
 
 // TestProperty_MultiTenantConstructors_NeverPanic verifies the nil-safety
 // invariant: no constructor or predicate ever panics for any combination of
-// nil/non-nil pointers. quick.Check generates bool values that drive whether
-// each pointer parameter is nil or non-nil.
-//
-// This guards against nil-pointer dereferences in constructors and predicates
-// when callers pass incomplete configurations (e.g., enabled=true but
-// pgManager=nil, or all parameters nil).
+// nil/non-nil pointers.
 func TestProperty_MultiTenantConstructors_NeverPanic(t *testing.T) {
 	t.Parallel()
 
@@ -304,9 +160,6 @@ func TestProperty_MultiTenantConstructors_NeverPanic(t *testing.T) {
 		}
 
 		// Property: NewBalanceSyncWorkerMT must never panic.
-		// Note: logger is always non-nil because the base constructor calls
-		// logger methods; passing nil logger would panic in production too,
-		// but that is a caller contract, not a multi-tenant invariant.
 		worker := NewBalanceSyncWorkerMT(logger, wUseCase, BalanceSyncConfig{}, workerEnabled, wTenantCache, wPGManager, "transaction")
 		if worker == nil {
 			return false
